Possible solution for sending data to pages from a layout
