+++ services/admin/src/routes/object/[prefix]/edit/[noid].svelte
@@ -1,22 +1,3 @@
-<script context="module" lang="ts">
-  import type { Load } from "@sveltejs/kit";
-  export const load: Load = async ({ page, fetch }) => {
-    const id = [
-      page.params["prefix"] as string,
-      page.params["noid"] as string,
-    ].join("/");
-    const response = await fetch(`/object/${id}.json`);
-    const json = await response.json();
-
-    if (response.ok) {
-      const object = json.object as AccessObject;
-      return { props: { object } };
-    } else {
-      return { status: response.status, error: new Error(json.error) };
-    }
-  };
-</script>
-
 <script lang="ts">
   import type { AccessObject } from "@crkn-rcdr/access-data";
   import { isCanvasManifest, isCollection } from "@crkn-rcdr/access-data";
@@ -25,31 +6,36 @@
   import SideMenuPage from "$lib/components/shared/SideMenuPage.svelte";
   import AccessObjectInfoEditor from "$lib/components/access-objects/AccessObjectInfoEditor.svelte";
   import EditorLayout from "$lib/components/access-objects/EditorLayout.svelte";
+  import { getContext } from "svelte";
 
-  export let object: AccessObject;
-  let type: string | undefined; //"collection" | "manifest" | "other";
-  import { page } from "$app/stores";
-  page.subscribe((page) => {
-    ({ type } = page.params);
-  });
+  let object: AccessObject;
+  const objectStore: SvelteStore<AccessObject> = getContext("objectStore");
+  $: object = $objectStore;
+  $: setDataModel(object);
+
+  let type: "collection" | "canvasManifest" | "other";
+  const typeStore: SvelteStore<"collection" | "canvasManifest" | "other"> =
+    getContext("typeStore");
+  $: type = $typeStore;
+  $: console.log(type);
 
   let rfdc: any; // Deep copies an object
-  let model: AccessObject; // | Canvas etc...
-
+  let model: AccessObject;
   async function setDataModel(object: AccessObject) {
     rfdc = (await import("rfdc")).default();
     model = rfdc(object) as AccessObject;
+    console.log(model);
   }
-
-  $: setDataModel(object);
 </script>
+
+<slot />
 
 {#if object && model}
   <!-- I couldn't have named slots in a regular __layout, not sure if there's a smarter way to make the layout more extensible-->
   <EditorLayout bind:object bind:model>
     <div slot="editor-menu">
       <!-- Not sure if this is any more performant than the isManifest check -->
-      {#if type === "manifest"}
+      {#if type === "canvasManifest"}
         <SideMenuPageListButton>General Info</SideMenuPageListButton>
         <SideMenuPageListButton>Content</SideMenuPageListButton>
       {:else if type === "collection"}
